### What?

- [x] Restore previous GET /allocations support for includes parameter
- [x] Restore previous GET /allocations default includes when includes parameter not included
- [x] Reinstate previous broken specs that were marked as pending temporarily

### Why?

- The previous change was made as part of an attempted hotfix yesterday. We observed slow performance on the `GET /allocations` endpoint and assumed this was due to returning additional relationships that were recently added.
- The original change made did not improve the observed performance on staging environment (which was caused by another bug returning all documents for a move without a profile)
- Reverting this change does not impact the current performance baseline against staging data (tested locally)

#### Before this PR:

```
[active_model_serializers] Rendered ActiveModel::Serializer::CollectionSerializer with ActiveModelSerializers::Adapter::JsonApi (1234.66ms)
Completed 200 OK in 1464ms (Views: 1261.5ms | ActiveRecord: 134.0ms | Allocations: 1066401)
```

#### After this PR:

```
[active_model_serializers] Rendered ActiveModel::Serializer::CollectionSerializer with ActiveModelSerializers::Adapter::JsonApi (1218.58ms)
Completed 200 OK in 1335ms (Views: 1174.8ms | ActiveRecord: 116.4ms | Allocations: 1055574)
```

### Deployment risks (optional)

- Reverts production api to previous behaviour; this behaviour was present for a few days prior to the unrelated document bug being introduced so is safe to reinstate. Tested against a local copy of staging data and performance of the endpoint with no filters specified remains the same with this change.